+++ dart/tool/dart/generate_dart.dart
@@ -70,23 +70,7 @@ final String _implCode = r'''
       if (json['id'] == null && json['method'] == 'streamNotify') {
         Map params = json['params'];
         String streamId = params['streamId'];
-
-        // TODO: These could be generated from a list.
-        if (streamId == 'VM') {
-          _vmController.add(_createObject(params['event']));
-        } else if (streamId == 'Isolate') {
-          _isolateController.add(_createObject(params['event']));
-        } else if (streamId == 'Debug') {
-          _debugController.add(_createObject(params['event']));
-        } else if (streamId == 'GC') {
-          _gcController.add(_createObject(params['event']));
-        } else if (streamId == 'Stdout') {
-          _stdoutController.add(_createObject(params['event']));
-        } else if (streamId == 'Stderr') {
-          _stderrController.add(_createObject(params['event']));
-        } else {
-          _log.warning('unknown streamId: ${streamId}');
-        }
+        _getEventController(streamId).add(_createObject(params['event']));
       } else if (json['id'] != null) {
         Completer completer = _completers.remove(json['id']);
 
@@ -237,6 +221,9 @@ class Api extends Member with ApiParseUtil {
 /// @optional
 const String optional = 'optional';
 
+/// @unstable
+const String unstable = 'unstable';
+
 /// Decode a string in Base64 encoding into the equivalent non-encoded string.
 /// This is useful for handling the results of the Stdout or Stderr events.
 String decodeBase64(String str) => new String.fromCharCodes(BASE64.decode(str));
@@ -259,13 +246,6 @@ Object _createObject(dynamic json) {
   }
 }
 
-String _printEnum(Object obj) {
-  if (obj == null) return null;
-  String str = obj.toString();
-  int index = str.indexOf('.');
-  return str.substring(index + 1);
-}
-
 ''');
     gen.writeln();
     gen.write('Map<String, Function> _typeFactories = {');
@@ -281,37 +261,52 @@ String _printEnum(Object obj) {
     gen.writeStatement('int _id = 0;');
     gen.writeStatement('Map<String, Completer<Response>> _completers = {};');
     gen.writeStatement('Log _log;');
-    gen.writeln();
-    gen.writeln("StreamController<String> _onSend = new StreamController.broadcast(sync: true);");
-    gen.writeln("StreamController<String> _onReceive = new StreamController.broadcast(sync: true);");
-    gen.writeln();
-    gen.writeln("StreamController<Event> _vmController = new StreamController.broadcast();");
-    gen.writeln("StreamController<Event> _isolateController = new StreamController.broadcast();");
-    gen.writeln("StreamController<Event> _debugController = new StreamController.broadcast();");
-    gen.writeln("StreamController<Event> _gcController = new StreamController.broadcast();");
-    gen.writeln("StreamController<Event> _stdoutController = new StreamController.broadcast();");
-    gen.writeln("StreamController<Event> _stderrController = new StreamController.broadcast();");
-    gen.writeln();
-    gen.writeStatement(
-        'VmService(Stream<String> inStream, void writeMessage(String message), {Log log}) {');
-    gen.writeStatement('_streamSub = inStream.listen(_processMessage);');
-    gen.writeStatement('_writeMessage = writeMessage;');
-    gen.writeStatement('_log = log == null ? new _NullLog() : log;');
-    gen.writeln('}');
-    gen.writeln();
-    gen.writeln("// VMUpdate");
-    gen.writeln("Stream<Event> get onVMEvent => _vmController.stream;");
-    gen.writeln("// IsolateStart, IsolateRunnable, IsolateExit, IsolateUpdate");
-    gen.writeln("Stream<Event> get onIsolateEvent => _isolateController.stream;");
-    gen.writeln("// PauseStart, PauseExit, PauseBreakpoint, PauseInterrupted, PauseException,");
-    gen.writeln("// Resume, BreakpointAdded, BreakpointResolved, BreakpointRemoved, Inspect");
-    gen.writeln("Stream<Event> get onDebugEvent => _debugController.stream;");
-    gen.writeln("// GC");
-    gen.writeln("Stream<Event> get onGCEvent => _gcController.stream;");
-    gen.writeln("// WriteEvent");
-    gen.writeln("Stream<Event> get onStdoutEvent => _stdoutController.stream;");
-    gen.writeln("// WriteEvent");
-    gen.writeln("Stream<Event> get onStderrEvent => _stderrController.stream;");
+    gen.writeln('''
+
+StreamController<String> _onSend = new StreamController.broadcast(sync: true);
+StreamController<String> _onReceive = new StreamController.broadcast(sync: true);
+
+Map<String, StreamController<Event>> _eventControllers = {};
+
+StreamController<Event> _getEventController(String eventName) {
+  StreamController<Event> controller = _eventControllers[eventName];
+  if (controller == null) {
+    controller = new StreamController.broadcast();
+    _eventControllers[eventName] = controller;
+  }
+  return controller;
+}
+
+VmService(Stream<String> inStream, void writeMessage(String message), {Log log}) {
+  _streamSub = inStream.listen(_processMessage);
+  _writeMessage = writeMessage;
+  _log = log == null ? new _NullLog() : log;
+}
+
+// VMUpdate
+Stream<Event> get onVMEvent => _getEventController('VM').stream;
+
+// IsolateStart, IsolateRunnable, IsolateExit, IsolateUpdate
+Stream<Event> get onIsolateEvent => _getEventController('Isolate').stream;
+
+// PauseStart, PauseExit, PauseBreakpoint, PauseInterrupted, PauseException,
+// Resume, BreakpointAdded, BreakpointResolved, BreakpointRemoved, Inspect
+Stream<Event> get onDebugEvent => _getEventController('Debug').stream;
+
+// GC
+Stream<Event> get onGCEvent => _getEventController('GC').stream;
+
+// WriteEvent
+Stream<Event> get onStdoutEvent => _getEventController('Stdout').stream;
+
+// WriteEvent
+Stream<Event> get onStderrEvent => _getEventController('Stderr').stream;
+
+// Listen for a specific event name.
+Stream<Event> onEvent(String streamName) => _getEventController('streamName').stream;
+
+''');
+
     gen.writeln();
     methods.forEach((m) => m.generate(gen));
     gen.out(_implCode);
@@ -365,11 +360,12 @@ class Method extends Member {
       gen.write('Future<${returnType.name}> ${name}(');
       bool startedOptional = false;
       gen.write(args.map((MethodArg arg) {
+        String typeName = api.isEnumName(arg.type) ? '/*${arg.type}*/ String' : arg.paramType;
         if (arg.optional && !startedOptional) {
           startedOptional = true;
-          return '{${arg.paramType} ${arg.name}';
+          return '{${typeName} ${arg.name}';
         } else {
-          return '${arg.paramType} ${arg.name}';
+          return '${typeName} ${arg.name}';
         }
       }).join(', '));
       if (startedOptional) gen.write('}');
@@ -384,9 +380,6 @@ class Method extends Member {
         gen.writeln('};');
         args.where((MethodArg a) => a.optional).forEach((MethodArg arg) {
           String valueRef = arg.name;
-          if (api.isEnumName(arg.type)) {
-            valueRef = '_printEnum(${arg.name})';
-          }
           gen.writeln("if (${arg.name} != null) m['${arg.name}'] = ${valueRef};");
         });
         gen.writeStatement("return _call('${name}', m);");
@@ -395,11 +388,7 @@ class Method extends Member {
         gen.writeStatement('{');
         gen.write("return _call('${name}', {");
         gen.write(args.map((MethodArg arg) {
-          if (api.isEnumName(arg.type)) {
-            return "'${arg.name}': _printEnum(${arg.name})";
-          } else {
-            return "'${arg.name}': ${arg.name}";
-          }
+          return "'${arg.name}': ${arg.name}";
         }).join(', '));
         gen.writeStatement('});');
         gen.writeStatement('}');
@@ -556,21 +545,21 @@ class Type extends Member {
     // ctors
     gen.writeln('${name}();');
     gen.writeln();
-    
+
     String superCall = superName == null ? '' : ": super._fromJson(json) ";
     gen.writeln('${name}._fromJson(Map json) ${superCall}{');
     fields.forEach((TypeField field) {
-      if (field.type.isSimple) {
+      if (field.type.isSimple || field.type.isEnum) {
         gen.write("${field.generatableName} = json['${field.name}']");
         if (field.defaultValue != null) {
           gen.write(' ?? ${field.defaultValue}');
         }
         gen.writeln(';');
-      } else if (field.type.isEnum) {
-        // Parse the enum.
-        String enumTypeName = field.type.types.first.name;
-        gen.writeln(
-          "${field.generatableName} = _parse${enumTypeName}[json['${field.name}']];");
+      // } else if (field.type.isEnum) {
+      //   // Parse the enum.
+      //   String enumTypeName = field.type.types.first.name;
+      //   gen.writeln(
+      //     "${field.generatableName} = _parse${enumTypeName}[json['${field.name}']];");
       } else if (field.type.isArray) {
         TypeRef fieldType = field.type.types.first;
         gen.writeln("${field.generatableName} = _createObject(json['${field.name}']) "
@@ -654,7 +643,8 @@ class TypeField extends Member {
   void generate(DartGenerator gen) {
     if (docs.isNotEmpty) gen.writeDocs(docs);
     if (optional) gen.write('@optional ');
-    gen.writeStatement('${type.name} ${generatableName};');
+    String typeName = api.isEnumName(type.name) ? '/*${type.name}*/ String' : type.name;
+    gen.writeStatement('${typeName} ${generatableName};');
     if (parent.fields.any((field) => field.hasDocs)) gen.writeln();
   }
 }
@@ -669,19 +659,17 @@ class Enum extends Member {
     _parse(new Tokenizer(definition).tokenize());
   }
 
+  String get prefix =>
+    name.endsWith('Kind') ? name.substring(0, name.length - 4) : name;
+
   void generate(DartGenerator gen) {
     gen.writeln();
     if (docs != null) gen.writeDocs(docs);
-    gen.writeStatement('enum ${name} {');
+    gen.writeStatement('class ${name} {');
+    gen.writeStatement('${name}._();');
+    gen.writeln();
     enums.forEach((e) => e.generate(gen));
     gen.writeStatement('}');
-
-    gen.writeln();
-    gen.writeStatement('Map<String, ${name}> _parse${name} = {');
-    gen.writeStatement(enums.map((EnumValue e) {
-      return "'${e.name}': ${e.parent.name}.${e.name}";
-    }).join(', '));
-    gen.writeStatement('};');
   }
 
   void _parse(Token token) {
@@ -700,9 +688,7 @@ class EnumValue extends Member {
 
   void generate(DartGenerator gen) {
     if (docs != null) gen.writeDocs(docs);
-    String suffix = ',';
-    if (parent.enums.last == this) suffix = '';
-    gen.writeStatement("${name}${suffix}");
+    gen.writeStatement("static const String k${name} = '${name}';");
   }
 }